Break building of root EntityField out of IR.OperationBuilder
So it can be reused by IR.FragmentBuilder
